+++ influxdb/influxdb.sh
@@ -166,31 +166,31 @@ fi
 # Check config variables are set
 #
 if [ -z \${DATACENTER+x} ]; then
-    echo 'DATACENTER is unset - see documentation to configure this flavour with the datacenter name. All parameters are mandatory.'
+    echo 'DATACENTER is unset - see documentation to configure this flavour with the datacenter name. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${NODENAME+x} ];
 then
-    echo 'NODENAME is unset - see documentation to configure this flavour with a name for this node. All parameters are mandatory.'
+    echo 'NODENAME is unset - see documentation to configure this flavour with a name for this node. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${CONSULSERVERS+x} ]; then
-    echo 'CONSULSERVERS is unset - please pass in one or more correctly-quoted, comma-separated addresses for consul peer IPs. Refer to documentation All parameters are mandatory.'
+    echo 'CONSULSERVERS is unset - please pass in one or more correctly-quoted, comma-separated addresses for consul peer IPs. Refer to documentation. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${IP+x} ]; then
-    echo 'IP is unset - see documentation to configure this flavour for an IP address. All parameters are mandatory.'
+    echo 'IP is unset - see documentation to configure this flavour for an IP address. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${VAULTSERVER+x} ];
 then
-    echo 'VAULTSERVER is unset - see documentation to set the vault server IP address. This is required to obtain certificates. All parameters are mandatory.'
+    echo 'VAULTSERVER is unset - see documentation to set the vault server IP address. This is required to obtain certificates. This parameter is mandatory.'
     exit 1
 fi
 # we need a token from the vault server
 if [ -z \${VAULTTOKEN+x} ];
 then
-    echo 'VAULTTOKEN is unset - a vault token is required to obtain certificates. Refer to documentation. All parameters are mandatory.'
+    echo 'VAULTTOKEN is unset - a vault token is required to obtain certificates. Refer to documentation. This parameter is mandatory.'
     exit 1
 fi
 # GOSSIPKEY is a 32 byte, Base64 encoded key generated with consul keygen for the consul flavour.
@@ -198,25 +198,25 @@ fi
 # We'll re-use the one from the consul flavour
 if [ -z \${GOSSIPKEY+x} ];
 then
-    echo 'GOSSIPKEY is unset - please provide a 32 byte base64 key from the (consul keygen key) command. All parameters are mandatory.'
+    echo 'GOSSIPKEY is unset - please provide a 32 byte base64 key from the (consul keygen key) command. This parameter is mandatory.'
     exit 1
 fi
 # optional logging to remote syslog server
 if [ -z \${REMOTELOG+x} ];
 then
-    echo 'REMOTELOG is unset - please provide the IP address of a loki server, or set a null value. All parameters are mandatory.'
-    exit 1
+    echo 'REMOTELOG is unset - please provide the IP address of a loki server, or set a null value. This parameter is optional.'
+    REMOTELOG=\"null\"
 fi
 # sftpuser credentials
 if [ -z \${SFTPUSER+x} ];
 then
-    echo 'SFTPUSER is unset - please provide a username to use for the SFTP user on the vault leader. All parameters are mandatory.'
+    echo 'SFTPUSER is unset - please provide a username to use for the SFTP user on the vault leader. This parameter is mandatory.'
     exit 1
 fi
 # sftpuser password
 if [ -z \${SFTPPASS+x} ];
 then
-    echo 'SFTPPASS is unset - please provide a password for the SFTP user on the vault leader. All parameters are mandatory.'
+    echo 'SFTPPASS is unset - please provide a password for the SFTP user on the vault leader. This parameter is mandatory.'
     exit 1
 fi