+++ postgresql-patroni/postgresql-patroni.sh
@@ -218,53 +218,53 @@ fi
 #
 if [ -z \${DATACENTER+x} ];
 then
-    echo 'DATACENTER is unset - see documentation to configure this flavour with the datacenter name. All parameters are mandatory.'
+    echo 'DATACENTER is unset - see documentation to configure this flavour with the datacenter name. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${CONSULSERVERS+x} ];
 then
-    echo 'CONSULSERVERS is unset - please pass in one or more correctly-quoted, comma-separated addresses for consul peer IPs. Refer to documentation All parameters are mandatory.'
+    echo 'CONSULSERVERS is unset - please pass in one or more correctly-quoted, comma-separated addresses for consul peer IPs. Refer to documentation. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${NODENAME+x} ];
 then
-    echo 'NODENAME is unset - see documentation to configure this flavour with a name for this node. All parameters are mandatory.'
+    echo 'NODENAME is unset - see documentation to configure this flavour with a name for this node. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${IP+x} ];
 then
-    echo 'IP is unset - see documentation to configure this flavour for an IP address. All parameters are mandatory.'
+    echo 'IP is unset - see documentation to configure this flavour for an IP address. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${SERVICETAG+x} ];
 then
-    echo 'SERVICETAG is unset - please set a service tag of master, replica or standby-leader. Refer to documentation. All parameters are mandatory.'
+    echo 'SERVICETAG is unset - please set a service tag of master, replica or standby-leader. Refer to documentation. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${ADMPASS+x} ];
 then
-    echo 'ADMPASS is unset - please set the password for the admin user. All parameters are mandatory.'
+    echo 'ADMPASS is unset - please set the password for the admin user. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${KEKPASS+x} ];
 then
-    echo 'KEKPASS is unset - please set the password for the superuser postgres user. All parameters are mandatory.'
+    echo 'KEKPASS is unset - please set the password for the superuser postgres user. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${REPPASS+x} ];
 then
-    echo 'REPPASS is unset - please set the replicator user password for replication. All parameters are mandatory.'
+    echo 'REPPASS is unset - please set the replicator user password for replication. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${VAULTSERVER+x} ];
 then
-    echo 'VAULTSERVER is unset - see documentation to set the vault server IP address. This is required to obtain certificates. All parameters are mandatory.'
+    echo 'VAULTSERVER is unset - see documentation to set the vault server IP address. This is required to obtain certificates. This parameter is mandatory.'
     exit 1
 fi
 # we need a token from the vault server
 if [ -z \${VAULTTOKEN+x} ];
 then
-    echo 'VAULTTOKEN is unset - a vault token is required to obtain certificates. Refer to documentation. All parameters are mandatory.'
+    echo 'VAULTTOKEN is unset - a vault token is required to obtain certificates. Refer to documentation. This parameter is mandatory.'
     exit 1
 fi
 # GOSSIPKEY is a 32 byte, Base64 encoded key generated with consul keygen for the consul flavour.
@@ -272,25 +272,25 @@ fi
 # We'll re-use the one from the consul flavour
 if [ -z \${GOSSIPKEY+x} ];
 then
-    echo 'GOSSIPKEY is unset - please provide a 32 byte base64 key from the (consul keygen key) command. All parameters are mandatory.'
+    echo 'GOSSIPKEY is unset - please provide a 32 byte base64 key from the (consul keygen key) command. This parameter is mandatory.'
     exit 1
 fi
 # optional logging to remote syslog server
 if [ -z \${REMOTELOG+x} ];
 then
-    echo 'REMOTELOG is unset - please provide the IP address of a loki server, or set a null value. All parameters are mandatory.'
-    exit 1
+    echo 'REMOTELOG is unset - please provide the IP address of a loki server. Defaulting to \"null\" if not set. This parameter is optional.'
+    REMOTELOG=\"null\"
 fi
 # sftpuser credentials
 if [ -z \${SFTPUSER+x} ];
 then
-    echo 'SFTPUSER is unset - please provide a username to use for the SFTP user on the vault leader. All parameters are mandatory.'
+    echo 'SFTPUSER is unset - please provide a username to use for the SFTP user on the vault leader. This parameter is mandatory.'
     exit 1
 fi
 # sftpuser password
 if [ -z \${SFTPPASS+x} ];
 then
-    echo 'SFTPPASS is unset - please provide a password for the SFTP user on the vault leader. All parameters are mandatory.'
+    echo 'SFTPPASS is unset - please provide a password for the SFTP user on the vault leader. This parameter is mandatory.'
     exit 1
 fi