+++ prometheus/prometheus.sh
@@ -175,31 +175,31 @@ fi
 # Check config variables are set
 #
 if [ -z \${DATACENTER+x} ]; then
-    echo 'DATACENTER is unset - see documentation to configure this flavour with the datacenter name. All parameters are mandatory.'
+    echo 'DATACENTER is unset - see documentation to configure this flavour with the datacenter name. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${NODENAME+x} ];
 then
-    echo 'NODENAME is unset - see documentation to configure this flavour with a name for this node. All parameters are mandatory.'
+    echo 'NODENAME is unset - see documentation to configure this flavour with a name for this node. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${CONSULSERVERS+x} ]; then
-    echo 'CONSULSERVERS is unset - please pass in one or more correctly-quoted, comma-separated addresses for consul peer IPs. Refer to documentation All parameters are mandatory.'
+    echo 'CONSULSERVERS is unset - please pass in one or more correctly-quoted, comma-separated addresses for consul peer IPs. Refer to documentation. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${IP+x} ]; then
-    echo 'IP is unset - see documentation to configure this flavour for an IP address. All parameters are mandatory.'
+    echo 'IP is unset - see documentation to configure this flavour for an IP address. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${VAULTSERVER+x} ];
 then
-    echo 'VAULTSERVER is unset - see documentation to set the vault server IP address. This is required to obtain certificates. All parameters are mandatory.'
+    echo 'VAULTSERVER is unset - see documentation to set the vault server IP address. This is required to obtain certificates. This parameter is mandatory.'
     exit 1
 fi
 # we need a token from the vault server
 if [ -z \${VAULTTOKEN+x} ];
 then
-    echo 'VAULTTOKEN is unset - a vault token is required to obtain certificates. Refer to documentation. All parameters are mandatory.'
+    echo 'VAULTTOKEN is unset - a vault token is required to obtain certificates. Refer to documentation. This parameter is mandatory.'
     exit 1
 fi
 # GOSSIPKEY is a 32 byte, Base64 encoded key generated with consul keygen for the consul flavour.
@@ -207,68 +207,68 @@ fi
 # We'll re-use the one from the consul flavour
 if [ -z \${GOSSIPKEY+x} ];
 then
-    echo 'GOSSIPKEY is unset - please provide a 32 byte base64 key from the (consul keygen key) command. All parameters are mandatory.'
+    echo 'GOSSIPKEY is unset - please provide a 32 byte base64 key from the (consul keygen key) command. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${SCRAPECONSUL+x} ];
 then
-    echo 'SCRAPECONSUL is unset - please include a list of special quoted IP:Port for consul servers to scrape. All parameters are mandatory.'
+    echo 'SCRAPECONSUL is unset - please include a list of special quoted IP:Port for consul servers to scrape. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${SCRAPENOMAD+x} ];
 then
-    echo 'SCRAPENOMAD is unset - please include a list of special quoted IP:Port for nomad servers to scrape. All parameters are mandatory.'
+    echo 'SCRAPENOMAD is unset - please include a list of special quoted IP:Port for nomad servers to scrape. This parameter is mandatory.'
     exit 1
 fi
 if [ -z \${SCRAPEDATABASE+x} ];
 then
-    echo 'SCRAPEDATABASE is unset - please include a list of special quoted IP:Port for postgresql servers to scrape. All parameters are mandatory.'
+    echo 'SCRAPEDATABASE is unset - please include a list of special quoted IP:Port for postgresql servers to scrape. This parameter is mandatory.'
     exit 1
 fi
 
 # optional logging to remote syslog server
 if [ -z \${REMOTELOG+x} ];
 then
-    echo 'REMOTELOG is unset - please provide the IP address of a loki server, or set a null value. All parameters are mandatory.'
-    exit 1
+    echo 'REMOTELOG is unset - please provide the IP address of a loki server. Defaulting to \"null\" if unset. This parameter is optional.'
+    REMOTELOG=\"null\"
 fi
 # sftpuser credentials
 if [ -z \${SFTPUSER+x} ];
 then
-    echo 'SFTPUSER is unset - please provide a username to use for the SFTP user on the vault leader. All parameters are mandatory.'
+    echo 'SFTPUSER is unset - please provide a username to use for the SFTP user on the vault leader. This parameter is mandatory.'
     exit 1
 fi
 # sftpuser password
 if [ -z \${SFTPPASS+x} ];
 then
-    echo 'SFTPPASS is unset - please provide a password for the SFTP user on the vault leader. All parameters are mandatory.'
+    echo 'SFTPPASS is unset - please provide a password for the SFTP user on the vault leader. This parameter is mandatory.'
     exit 1
 fi
 # smtp settings for prometheus alertmanager config
 if [ -z \${SMTPHOSTPORT+x} ];
 then
-    echo 'SMTPHOSTPORT is unset - please provide an SMTP host:port paramater. If nothing then localhost:25 will do. All parameters are mandatory.'
-    exit 1
+    echo 'SMTPHOSTPORT is unset - please provide an SMTP host:port paramater. Defaulting to localhost:25 if unset. This parameter is optional.'
+    SMTPHOSTPORT=\"localhost:25\"
 fi
 if [ -z \${SMTPFROM+x} ];
 then
-    echo 'SMTPFROM is unset - please provide the from address for alert notifications. You can use a default of alertmanager@localhost if nothing else. All parameters are mandatory.'
-    exit 1
+    echo 'SMTPFROM is unset - please provide the from address for alert notifications. Defaulting to alertmanager@localhost. This parameter is optional.'
+    SMTPFROM=\"alertmanager@localhost\"
 fi
 if [ -z \${SMTPUSER+x} ];
 then
-    echo 'SMTPUSER is unset - please provide the smtp username for authenticated smtp connections. Set an empty value if none. All parameters are mandatory.'
-    exit 1
+    echo 'SMTPUSER is unset - please provide the smtp username for authenticated smtp connections. Set an empty value if none. Defaulting to none.'
+    SMTPUSER=\"\"
 fi
 if [ -z \${SMTPPASS+x} ];
 then
-    echo 'SMTPPASS is unset - please provide the smtp password for authenticated smtp connections. Set an empty value if none. All parameters are mandatory.'
-    exit 1
+    echo 'SMTPPASS is unset - please provide the smtp password for authenticated smtp connections. Set an empty value if none. Defaulting to none.'
+    SMTPPASS=\"\"
 fi
 if [ -z \${ALERTADDRESS+x} ];
 then
-    echo 'ALERTADDRESS is unset - please provide an email address to send alerts to. Set a value like notify@localhost if none. All parameters are mandatory.'
-    exit 1
+    echo 'ALERTADDRESS is unset - please provide an email address to send alerts to. Defaulting to notify@localhost. This parameter is optional.'
+    ALERTADDRESS=\"notify@localhost\"
 fi